# feat: 35% speed up for writing vcf files