--- a/vcf/filter_test.go
+++ b/vcf/filter_test.go
@@ -54,3 +54,41 @@ func TestIsBiallelic(t *testing.T) {
 		}
 	}
 }
+
+func TestIsSubstitution(t *testing.T) {
+	tests := []struct {
+		vcf      Vcf
+		expected bool
+	}{
+		{Vcf{Ref: "A", Alt: []string{"C"}}, true},
+		{Vcf{Ref: "AT", Alt: []string{"C"}}, false},
+		{Vcf{Ref: "A", Alt: []string{"CG"}}, false},
+	}
+
+	for _, test := range tests {
+		result := IsSubstitution(test.vcf)
+		if result != test.expected {
+			t.Errorf("Error: IsSubstitution(%v) = %v; want %v", test.vcf, result, test.expected)
+		}
+	}
+}
+
+func TestIsSegregating(t *testing.T) {
+	tests := []struct {
+		vcf      Vcf
+		expected bool
+	}{
+		{Vcf{Samples: []Sample{}}, false},
+		{Vcf{Samples: []Sample{{Alleles: []int16{0, 0}}}}, false},
+		{Vcf{Samples: []Sample{{Alleles: []int16{0, 1}}}}, true},
+		{Vcf{Samples: []Sample{{Alleles: []int16{1, 1}}}}, false},
+		{Vcf{Samples: []Sample{{Alleles: []int16{1}}, {Alleles: []int16{2}}}}, true},
+	}
+
+	for _, test := range tests {
+		result := IsSegregating(test.vcf)
+		if result != test.expected {
+			t.Errorf("Error: IsSegregating(%v) = %v; want %v", test.vcf, result, test.expected)
+		}
+	}
+}