--- a/vcf/filter_test.go
+++ b/vcf/filter_test.go
@@ -92,3 +92,110 @@ func TestIsSegregating(t *testing.T) {
 		}
 	}
 }
+
+func TestIsRefWeakAltStrong(t *testing.T) {
+	tests := []struct {
+		vcf      Vcf
+		expected bool
+	}{
+
+		{Vcf{Ref: "T", Alt: []string{"C"}}, true},
+		{Vcf{Ref: "T", Alt: []string{"G"}}, true},
+		{Vcf{Ref: "A", Alt: []string{"C", "G"}}, false},
+	}
+
+	for _, test := range tests {
+		result := IsRefWeakAltStrong(test.vcf)
+		if result != test.expected {
+			t.Errorf("IsRefWeakAltStrong(%v) = %v; want %v", test.vcf, result, test.expected)
+		}
+	}
+}
+
+func TestRefStrongAltWeak(t *testing.T) {
+	tests := []struct {
+		vcf      Vcf
+		expected bool
+	}{
+		{Vcf{Ref: "C", Alt: []string{"A"}}, true},
+		{Vcf{Ref: "C", Alt: []string{"T"}}, true},
+		{Vcf{Ref: "G", Alt: []string{"C"}}, false},
+		{Vcf{Ref: "C", Alt: []string{"A", "T"}}, false},
+	}
+
+	for _, test := range tests {
+		result := IsRefStrongAltWeak(test.vcf)
+		if result != test.expected {
+			t.Errorf("Error: IsRefStrongAltWeak(%v) = %v; want %v", test.vcf, result, test.expected)
+		}
+	}
+}
+
+func TestIsNotRefStrongAltWeak(t *testing.T) {
+	tests := []struct {
+		vcf      Vcf
+		expected bool
+	}{
+		{Vcf{Ref: "C", Alt: []string{"A"}}, false},
+		{Vcf{Ref: "G", Alt: []string{"C"}}, true},
+	}
+
+	for _, test := range tests {
+		result := IsNotRefStrongAltWeak(test.vcf)
+		if result != test.expected {
+			t.Errorf("Error: IsNotRefStrongAltWeak(%v) = %v; want %v", test.vcf, result, test.expected)
+		}
+	}
+}
+func TestIsNotRefWeakAltStrong(t *testing.T) {
+	tests := []struct {
+		vcf      Vcf
+		expected bool
+	}{
+		{Vcf{Ref: "C", Alt: []string{"A"}}, true},
+		{Vcf{Ref: "A", Alt: []string{"C"}}, false},
+	}
+
+	for _, test := range tests {
+		result := IsNotRefWeakAltStrong(test.vcf)
+		if result != test.expected {
+			t.Errorf("Error: IsNotRefWeakAltStrong(%v) = %v; want %v", test.vcf, result, test.expected)
+		}
+	}
+}
+
+func TestIsWeakToStrongOrStrongToWeak(t *testing.T) {
+	tests := []struct {
+		vcf      Vcf
+		expected bool
+	}{
+		{Vcf{Ref: "G", Alt: []string{"T"}}, true},
+		{Vcf{Ref: "G", Alt: []string{"C"}}, false},
+	}
+
+	for _, test := range tests {
+		result := IsWeakToStrongOrStrongToWeak(test.vcf)
+		if result != test.expected {
+			t.Errorf("Error: IsWeakToStrongOrStrongToWeak(%v) = %v; want %v", test.vcf, result, test.expected)
+		}
+	}
+}
+
+func TestIsNotWeakToStrongOrStrongToWeak(t *testing.T) {
+	tests := []struct {
+		vcf      Vcf
+		expected bool
+	}{
+		{Vcf{Ref: "A", Alt: []string{"C"}}, false},
+		{Vcf{Ref: "A", Alt: []string{"G"}}, false},
+		{Vcf{Ref: "T", Alt: []string{"C"}}, false},
+		{Vcf{Ref: "T", Alt: []string{"G"}}, false},
+	}
+
+	for _, test := range tests {
+		result := IsNotWeakToStrongOrStrongToWeak(test.vcf)
+		if result != test.expected {
+			t.Errorf("Error: IsNotWeakToStrongOrStrongToWeak(%v) = %v; want %v", test.vcf, result, test.expected)
+		}
+	}
+}