--- a/vcf/filter_test.go
+++ b/vcf/filter_test.go
@@ -1,6 +1,8 @@
 package vcf
 
-import "testing"
+import (
+	"testing"
+)
 
 func TestIsHeterozygous(t *testing.T) {
 	tests := []struct {