… vcf/io_test.go
… vcf/methods_test.go
… vcf/testdata/test.vcf.gz
… vcf/vcfTools_test.go
… vcf/vcf_test.go