# Description

- **~35%** improvements and optimizations for writing vcfs.
- when constructing strings, strings.Builder{} appears to be better performing when combining strings from multiple data fields using something like fmt.Sprintf("%s%n...").
- in the 'string.Builder{}' implementation, require a slightly more involved process to build out the final string, however the downstream performance gains is beneficial, especially when working with larger vcf files.


```
BenchmarkPreviousWrite-12                 4707           1229806 ns/op         20186 allocs/op
BenchmarkNewBetterWrite-12                6258            960211 ns/op          9274 allocs/op
```
- Key performance improvements to note: ~35% faster, and roughly 50% less memory allocation

### Testing
- At first glance, this pr may feel beefy but mostly contains tests for any modified code as well as test coverage in `filer.go`, most of first just check for conditions in vcf samples.

- More benchmarking:
```
edotau@eau-macBook-m3-pro:~/src/github.com/vertgenlab/gonomics/vcf$ go test -bench=. -benchtime=5s

BenchmarkWrite-12                   4707           1229806 ns/op          365237 B/op      20186 allocs/op
BenchmarkNewWrite-12               6274            955157 ns/op          963356 B/op       9274 allocs/op
BenchmarkWriteGzip-12               3270           1769856 ns/op         5903577 B/op      20230 allocs/op
BenchmarkNewWriteGzip-12            3499           1677263 ns/op         6501878 B/op       9318 allocs/op

PASS
ok      github.com/vertgenlab/gonomics/vcf      54.065s
```

### Checklist before requesting a review

- [x] I performed a self-review of my code
- [x] If it this a core feature, I added thorough tests
- [x] The command `go fmt` or `make clean` was used on all files included

<!-- ### Screenshots & Media
if relevant, add an screenshots, images or recordings -->

<!-- ### Edge cases / Breaking Changes / Known Issues
if relevant, document any edge cases, known issues, etc -->